requirements.txt
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
numpy>=1.26.0
opencv-python>=4.9.0.80
pillow>=10.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
tensorboard>=2.16.0
matplotlib>=3.8.0
rasterio>=1.3.9
shapely>=2.0.4
albumentations>=1.4.2
einops>=0.7.0
scipy>=1.11.0
psutil>=5.9.8
